feat: Implement CheckIOBlock, HeadCountBlock
1. I renamed the directory from search-stay to search-option-stay

2. I used the useComponentVisible to handle a click events in
 CheckIOBlock and HeadCountBlock.

3-1.  I implemented a StayBlock to deliver props to the child componets.
To avoid the component duplication.

3-2. I moved StayWrapper to SearchBlock, parent component.
To avoid the component duplication.

4. I modifed the shadow color in the buttons.

5. I removed the classname of the input-wrapper.
